added StartBuilding

+++ Messages/RpgMessage.cs
@@ -5,7 +5,7 @@ namespace ws_hero.Messages
         public string PlayerId { get; set; }
 
         public int Cid { get; set; }
-        public int ClientTime { get; set; }
+        public long ClientTime { get; set; }
 
         public RpgType RpgType { get; set; }
         public string Data { get; set; }
